requirements.txt
flask
mysql-connector-python==9.1.0